Fix namespaces after moving files to folders
