+++ IO/CsvReaderWriter.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.IO;
 
-namespace Payslip_Round_2
+namespace Payslip_Round_2.IO
 {
     public class CsvReaderWriter
     {
