+++ IO/UserInput.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Payslip_Round_2
+namespace Payslip_Round_2.IO
 {
     public static class UserInput
     {
