+++ TaxCalculator.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Dynamic;
 
 namespace Payslip_Round_2
 {
